Streamlit/st_requirements.txt

google-generativeai == 0.6.0
streamlit == 1.35.0
temp == 2020.7.2